modified the console view and updated team and users routes

+++ public/javascripts/controller/ConsoleController.js
@@ -7,7 +7,12 @@ var ConsoleController = App.controller('ConsoleController', function($scope,$doc
     $scope.players = [];
 
     $scope.showUserList = false;
+    $scope.tempUsername = "";
+    $scope.tempPassword = "";
     $scope.users = [];
+
+    $scope.showTeamList = false;
+    $scope.teams = [];
 
     $scope.loadPlayers = function(){
         var players = PlayerService.getAll().then(function(response){
@@ -34,16 +39,29 @@ var ConsoleController = App.controller('ConsoleController', function($scope,$doc
         });
     }
 
-    $scope.createUser = function(){
-        UserService.create().then(function(response){
-            $scope.players.push(response.data);
+    $scope.createUser = function(username, password){
+        UserService.create(username, password).then(function(response){
+            if(response.data.id){
+                $scope.users.push(response.data);
+            }else{
+                alert(response.data);
+                console.log(response.data);
+            }
         });
     }
 
     $scope.deleteUser = function(user){
-        UserService.deleteById(user.id).then(function(response){
-            var index = $scope.users.indexOf(user);
-            $scope.users.splice(index, 1);
+        if(user && user.id){
+            UserService.deleteById(user.id).then(function(response){
+                var index = $scope.users.indexOf(user);
+                $scope.users.splice(index, 1);
+            });
+        }
+    }
+
+    $scope.loadTeams = function(){
+        var teams = TeamService.getAll().then(function(response){
+            $scope.teams = response.data.teams;
         });
     }
 });
